pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = ""
authors = [
    {name = "Rafael Celente", email = "rafaelmcelente@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch==2.8",
    "pydantic>=2.11.5,<3.0.0",
    "pyyaml>=6.0.2,<7.0.0",
    "lightning>=2.5.1.post0,<3.0.0",
    "datasets>=2.20.0,<3.0.0",
    "transformers>=4.43.3,<5.0.0",
    "wandb>=0.21.1,<0.22.0",
    "matplotlib>=3.10.5,<4.0.0",
    "anyscale>=0.26.58",
    "einops>=0.8.2",
    "quack-kernels>=0.2.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
]
test = [
    "pytest>=8.4.0",
]

[dependency-groups]
dev = [
    "google-cloud-storage>=3.4.0",
    "matplotlib>=3.9.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/core"]
